refactor: extract grammar string scanner
